+++ buildSrc/src/main/kotlin/AppConfig.kt
@@ -4,6 +4,6 @@ object AppConfig {
     const val minSdkVersion = 24
     const val ndkVersion = "28.0.13004108"
 
-    const val versionCode = 43
-    const val versionName = "2.0.3-widescreen"
+    const val versionCode = 44
+    const val versionName = "2.0.4-widescreen"
 }